+++ CHANGELOG.md
@@ -6,6 +6,7 @@ HEAD
 
 * Fixed error `ambiguous overload for 'operator|'` (issue #1411)
 * Fixed `operator|(MemberProxy, JsonObject)` (issue #1415)
+* Allowed more than 32767 values in non-embedded mode (issue #1414)
 
 v6.17.0 (2020-10-19)
 -------